- Implemented a reasonable collision detection by using additional potential fields - Performance improvements by the use of temporary values - Bugfix: The movement distance per timestep was too high

+++ include/UnitGenes.h
@@ -168,7 +168,7 @@ public:
         initHash();
     }
 
-    int get(size_t const pos) const
+    int get(int const pos) const
     {
         return X[pos];
     }
